move staking to celo builder
