+++ src/coin/eth/transactionBuilder.ts
@@ -2,7 +2,7 @@ import { BaseCoin as CoinConfig } from '@bitgo/statics/dist/src/base';
 import BigNumber from 'bignumber.js';
 import { RLP } from 'ethers/utils';
 import * as Crypto from '../../utils/crypto';
-import { BaseTransaction, BaseTransactionBuilder, TransactionType, StakingOperationsTypes } from '../baseCoin';
+import { BaseTransaction, BaseTransactionBuilder, TransactionType } from '../baseCoin';
 import { BaseAddress, BaseKey } from '../baseCoin/iface';
 import { Transaction, TransferBuilder, Utils } from '../eth';
 import {
@@ -12,8 +12,6 @@ import {
   ParseTransactionError,
   SigningError,
 } from '../baseCoin/errors';
-import { StakingBuilder } from '../cgld/stakingBuilder';
-import { Staking } from '../cgld/staking';
 import { KeyPair } from './keyPair';
 import { Fee, SignatureParts, TxData } from './iface';
 import {
@@ -30,9 +28,9 @@ const DEFAULT_M = 3;
  * Ethereum transaction builder.
  */
 export class TransactionBuilder extends BaseTransactionBuilder {
+  protected _type: TransactionType;
   private _transaction: Transaction;
   private _sourceKeyPair: KeyPair;
-  private _type: TransactionType;
   private _chainId: number;
   private _counter: number;
   private _fee: Fee;
@@ -49,9 +47,6 @@ export class TransactionBuilder extends BaseTransactionBuilder {
   private _contractAddress: string;
   private _contractCounter: number;
 
-  // Staking specific parameters
-  private _stakingBuilder: StakingBuilder;
-
   /**
    * Public constructor.
    *
@@ -67,23 +62,7 @@ export class TransactionBuilder extends BaseTransactionBuilder {
 
   /** @inheritdoc */
   protected async buildImplementation(): Promise<BaseTransaction> {
-    let transactionData;
-    switch (this._type) {
-      case TransactionType.WalletInitialization:
-        transactionData = this.buildWalletInitializationTransaction();
-        break;
-      case TransactionType.Send:
-        transactionData = this.buildSendTransaction();
-        break;
-      case TransactionType.AddressInitialization:
-        transactionData = this.buildAddressInitializationTransaction();
-        break;
-      case TransactionType.Staking_Lock:
-        transactionData = this.buildLockStakeTransaction();
-        break;
-      default:
-        throw new BuildTransactionError('Unsupported transaction type');
-    }
+    const transactionData = this.getTransactionData();
 
     if (this._txSignature) {
       Object.assign(transactionData, this._txSignature);
@@ -97,6 +76,20 @@ export class TransactionBuilder extends BaseTransactionBuilder {
       await this.transaction.sign(this._sourceKeyPair);
     }
     return this.transaction;
+  }
+
+  protected getTransactionData(): TxData {
+    let transactionData;
+    switch (this._type) {
+      case TransactionType.WalletInitialization:
+        return this.buildWalletInitializationTransaction();
+      case TransactionType.Send:
+        return this.buildSendTransaction();
+      case TransactionType.AddressInitialization:
+        return this.buildAddressInitializationTransaction();
+      default:
+        throw new BuildTransactionError('Unsupported transaction type');
+    }
   }
 
   /** @inheritdoc */
@@ -315,7 +308,7 @@ export class TransactionBuilder extends BaseTransactionBuilder {
     this._sourceAddress = source;
   }
 
-  private buildBase(data: string): TxData {
+  protected buildBase(data: string): TxData {
     return {
       gasLimit: this._fee.gasLimit,
       gasPrice: this._fee.fee,
@@ -446,34 +439,6 @@ export class TransactionBuilder extends BaseTransactionBuilder {
     }
     return calculateForwarderAddress(this._contractAddress, this._contractCounter);
   }
-  //endregion
-
-  //region Stake methods
-  public lock(): StakingBuilder {
-    if (this._type !== TransactionType.Staking_Lock) {
-      throw new BuildTransactionError('Lock can only be set for Staking Lock transactions type');
-    }
-    if (!this._stakingBuilder) {
-      this._stakingBuilder = new StakingBuilder().type(StakingOperationsTypes.LOCK);
-    }
-    return this._stakingBuilder;
-  }
-
-  private getStaking(): Staking {
-    if (!this._stakingBuilder) {
-      throw new BuildTransactionError('No staking information set');
-    }
-    return this._stakingBuilder.build();
-  }
-
-  private buildLockStakeTransaction(): TxData {
-    const stake = this.getStaking();
-    const data = this.buildBase(stake.serialize());
-    data.to = stake.address;
-    data.value = stake.amount;
-    return data;
-  }
-
   //endregion
 
   /** @inheritdoc */
